Move flux coordinate test to test_equilibrium
