+++ tests/test_equilibrium.py
@@ -85,6 +85,7 @@ def test_magnetic_axis_guess(DummyStellarator):
 
 
 def test_compute_theta_coords(SOLOVEV):
+    """Test root finding for theta(theta*, lambda(theta))"""
 
     eq = EquilibriaFamily.load(load_from=str(SOLOVEV["desc_h5_path"]))[-1]
 
@@ -97,13 +98,35 @@ def test_compute_theta_coords(SOLOVEV):
     flux_coords = nodes.copy()
     flux_coords[:, 1] += coords["lambda"]
 
-    geom_coords = np.asarray(eq.compute_theta_coords(flux_coords))
+    geom_coords = np.array(eq.compute_theta_coords(flux_coords))
 
     # catch difference between 0 and 2*pi
     if geom_coords[0, 1] > np.pi:  # theta[0] = 0
         geom_coords[0, 1] = geom_coords[0, 1] - 2 * np.pi
 
     np.testing.assert_allclose(nodes, geom_coords, rtol=1e-5, atol=1e-5)
+
+
+def test_compute_flux_coords(SOLOVEV):
+    """Test root finding for (rho,theta,zeta) from (R,phi,Z)"""
+
+    eq = EquilibriaFamily.load(load_from=str(SOLOVEV["desc_h5_path"]))[-1]
+
+    rho = np.linspace(0.01, 0.99, 20)
+    theta = np.linspace(0, 2 * np.pi, 20, endpoint=False)
+    zeta = np.linspace(0, 2 * np.pi, 20, endpoint=False)
+
+    nodes = np.vstack([rho, theta, zeta]).T
+    coords = eq.compute_toroidal_coords(Grid(nodes, sort=False))
+    real_coords = np.vstack([coords["R"].flatten(), zeta, coords["Z"].flatten()]).T
+
+    flux_coords = np.array(eq.compute_flux_coords(real_coords))
+
+    # catch difference between 0 and 2*pi
+    if flux_coords[0, 1] > np.pi:  # theta[0] = 0
+        flux_coords[0, 1] = flux_coords[0, 1] - 2 * np.pi
+
+    np.testing.assert_allclose(nodes, flux_coords, rtol=1e-5, atol=1e-5)
 
 
 # can't test booz_xform because it can't be installed by Travis
